@aexmachina 2015-03-24
Hi there, thanks for that but I'm afraid it's not a solution. I have found some performance improvements that should fix your problem in v2.0.4. Let me know if you're still having issues.