@huehnerhose 2015-03-25
works perfectly now. Complete Bonsai render takes about 200-250ms :) Thanks!